refactor: replace type

+++ tool/substract.go
@@ -1,6 +1,6 @@
 package tool
 
 // It substract a value with b
-func Substract(a, b int) int {
+func Substract(a, b float64) float64 {
 	return a - b
 }
